Adds direct dispatching of call messages. #42553
- Checks if "DirectMessageDispatch" capacity is allowed to activate;
- Tests check that the dispatch duration does not grow linearly
  with the number of objects;
- ServiceObjects and RemoteObjects are now unique per object
  (identified using ptruid);

This last point implies that previously we were creating N RemoteObjects
for N receptions of (Any)Objects pointing to the same object.
This patch makes the receiving and sending code detect if a RemoteObject
and/or ServiceObject already exist and use it if it does.
Consequently, a memory usage reduction is expected.

Change-Id: I8e545bc640ccfcfdf67cad10326ec53c2d0554c7
Reviewed-on: http://gerrit.aldebaran.lan/92429
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: matthieu.paindavoine <matthieu.paindavoine@softbankrobotics.com>