Adds ObjectUid as universal ID of remote objects implemented using PtrUid.
This change removes the implication that we are exposing pointers
in qi::Object interface through qi::Object::ptruid().
We deprecate qi::Object::ptruid() and provide qi::Object::uid() as
a clearer replacement.
The implementation still relies on ObjectUid to have the interface
and be compatible with PtrUid but we can now separate the meaning
of an unique identifier of an object refered by qi::Object from an
universally unique pointer value that PtrUid represents,
allowing for ObjectUid to be implemented differently in the future
and to clarify the role of qi::Object::uid().

Change-Id: Ib91deb66ec1a6fc3c9e14d1ac679bd61d27cdcd2
Reviewed-on: http://gerrit.aldebaran.lan/92985
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>